requirements.txt
asgiref==3.2.5
attrs==19.3.0
autopep8==1.5
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.3
cryptography==2.8
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.4
django-cors-headers==3.2.1
django-coverage-plugin==1.8.0
django-environ==0.4.5
django-extensions==2.2.8
django-filter==2.2.0
django-import-export==2.0.2
django-prometheus==2.0.0
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
entrypoints==0.3
et-xmlfile==1.0.1
flake8==3.7.9
idna==2.9
importlib-metadata==1.5.0
isort==4.3.21
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.3
packaging==20.3
Pillow==7.0.0
pluggy==0.13.1
prometheus-client==0.7.1
psycopg2-binary==2.8.4
py==1.8.1
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
PyJWT==1.7.1
pyparsing==2.4.6
pytest==5.3.5
pytest-django==3.8.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
tablib==1.1.0
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.8
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
